Prueba solucion error boton eliminar

+++ src/main/java/com/paula/services/BookService.java
@@ -1,6 +1,9 @@
 package com.paula.services;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -25,8 +28,15 @@ public class BookService {
         return bookRepository.findByUser(user);
     }
     
+	@Transactional
     public void deleteBookById(Integer bookId) {
-        bookRepository.deleteById(bookId);
+        
+        Optional<Book> book = bookRepository.findById(bookId);
+		if (book.isPresent()) {
+			bookRepository.deleteById(bookId);
+		} else {
+			throw new RuntimeException("¡Error! El libro con id " + bookId + " no se encuentra en la base de datos.");
+		}
     }
     
 }
